requirements.txt
python-dotenv~=0.20.0
Flask~=2.3.2
pip~=24.3.1